Fix type

+++ api.py
@@ -6,7 +6,7 @@ app = FastAPI()
 
 
 @app.get("/quadratic_equation")
-async def quadratic_equation(a: int, b: int, c: int):
+async def quadratic_equation(a: float, b: float, c: float):
     """Endpoint для вычесления квадратного уравнения"""
     massage = kv_ur(a, b, c)
     return {"Ответ": massage}
